Log distance to uint16 and add public neighbours calls (#371)
* Use uint16 instead of uint32 for discv5 log distance

* Make neighboursAtDistances and neighbours calls available

+++ eth/p2p/discoveryv5/dcli.nim
@@ -79,7 +79,7 @@ type
       distance* {.
         defaultValue: 255
         desc: "Distance parameter for the findNode message"
-        name: "distance" .}: uint32
+        name: "distance" .}: uint16
       # TODO: Order here matters as else the help message does not show all the
       # information, see: https://github.com/status-im/nim-confutils/issues/15
       findNodeTarget* {.
